requirements.txt
# National Revolution 1404 Telegram Bot
# Python dependencies

# Telegram Bot Framework (v20+)
python-telegram-bot==20.7

# Database
# SQLite is built-in, no additional packages needed

# Image/Video Processing (for metadata stripping)
# Note: Requires ffmpeg to be installed on system
# Install: https://ffmpeg.org/download.html

# OCR for Conduit verification (optional)
pytesseract>=0.3.10
Pillow>=10.4.0

# QR Code generation for certificates
qrcode[pil]>=7.4.0

# Web server for Mini App (optional, if self-hosting)
flask>=3.0.0
flask-cors>=4.0.0

# Security
cryptography>=42.0.0
pysqlcipher3>=1.2.0
python-dotenv>=1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
python-json-logger==2.0.7

# Optional: PostgreSQL support (instead of SQLite for production)
# Uncomment if using PostgreSQL:
# psycopg2-binary==2.9.9

# Optional: Redis for caching (advanced)
# redis==5.0.1

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
